feat: Better train sound & mod interaction framework

+++ src/main/java/createvvvfsim/VVVFSoundEngine.java
@@ -7,14 +7,14 @@ import javax.sound.sampled.AudioSystem;
 import javax.sound.sampled.LineUnavailableException;
 import javax.sound.sampled.SourceDataLine;
 public class VVVFSoundEngine{
-    public static float settings_amp=1f;
+    public static double settings_amp=1.0;
     public static boolean is_paused=false;
     public static final int sample_rate=44100;
     public static final int buffer_size=1<<12;
     private static final AudioFormat format=new AudioFormat(sample_rate,16,1,true,false);
     private static final Object mix_lock=new Object();
     private static final List<VVVFSoundGen> generators=new ArrayList<>();
-    private static final float[] mix_buffer=new float[buffer_size];
+    private static final double[] mix_buffer=new double[buffer_size];
     private static final byte[] out_buffer=new byte[buffer_size*2];
     private static final Thread thread=new Thread(VVVFSoundEngine::mixLoop);
     private static SourceDataLine dataline;
@@ -33,7 +33,7 @@ public class VVVFSoundEngine{
             is_paused=true;
         }
     }
-    public static void offPause(float volume){
+    public static void offPause(double volume){
         synchronized(mix_lock){
             settings_amp=volume;
             is_paused=false;
@@ -43,7 +43,7 @@ public class VVVFSoundEngine{
     private static void mixLoop(){
         while(true){
             synchronized(mix_lock){
-                while(generators.isEmpty() || is_paused || settings_amp<1e-3f){
+                while(generators.isEmpty() || is_paused || settings_amp<1e-3){
                     try{
                         mix_lock.wait();
                     }
@@ -53,11 +53,11 @@ public class VVVFSoundEngine{
                 }
             }
             synchronized(mix_lock){
-                Arrays.fill(mix_buffer,0f);
+                Arrays.fill(mix_buffer,0.0);
                 for(VVVFSoundGen gen:generators) gen.mixTo(mix_buffer);
             }
             for(int i=0;i<buffer_size;i++){
-                float clipped=Math.clamp(mix_buffer[i],-1f,1f)*settings_amp;
+                double clipped=Math.clamp(mix_buffer[i],-1.0,1.0)*settings_amp;
                 short sample=(short)(clipped*Short.MAX_VALUE);
                 out_buffer[i*2]=(byte)(sample&0xFF);
                 out_buffer[i*2+1]=(byte)((sample>>8)&0xFF);
