+++ src/main/java/vvvfsimulator/audiofilter/FFTConvolver.java
@@ -9,27 +9,27 @@ public class FFTConvolver {
     private int segSize;
     private int segCount;
     private int fftComplexSize;
-    private final List<float[]> segmentsRe = new ArrayList<>();
-    private final List<float[]> segmentsIm = new ArrayList<>();
-    private final List<float[]> segmentsIRRe = new ArrayList<>();
-    private final List<float[]> segmentsIRIm = new ArrayList<>();
-    private float[] fftBuffer = new float[0];
+    private final List<double[]> segmentsRe = new ArrayList<>();
+    private final List<double[]> segmentsIm = new ArrayList<>();
+    private final List<double[]> segmentsIRRe = new ArrayList<>();
+    private final List<double[]> segmentsIRIm = new ArrayList<>();
+    private double[] fftBuffer = new double[0];
     private final AudioFFT fft = new AudioFFT();
-    private float[] preMultipliedRe = new float[0];
-    private float[] preMultipliedIm = new float[0];
-    private float[] convRe = new float[0];
-    private float[] convIm = new float[0];
-    private float[] overlap = new float[0];
+    private double[] preMultipliedRe = new double[0];
+    private double[] preMultipliedIm = new double[0];
+    private double[] convRe = new double[0];
+    private double[] convIm = new double[0];
+    private double[] overlap = new double[0];
     private int current;
-    private float[] inputBuffer = new float[0];
+    private double[] inputBuffer = new double[0];
     private int inputBufferFill;
 
-    public boolean init(int blockSize, float[] ir, int irLen) {
+    public boolean init(int blockSize, double[] ir, int irLen) {
         reset();
         if (blockSize == 0) {
             return false;
         }
-        while (irLen > 0 && Math.abs(ir[irLen - 1]) < 0.000001f) {
+        while (irLen > 0 && Math.abs(ir[irLen - 1]) < 0.000001) {
             irLen--;
         }
         if (irLen == 0) {
@@ -41,16 +41,16 @@ public class FFTConvolver {
         this.fftComplexSize = AudioFFT.ComplexSize(segSize);
 
         fft.init(segSize);
-        fftBuffer = new float[segSize];
+        fftBuffer = new double[segSize];
 
         for (int i = 0; i < segCount; i++) {
-            segmentsRe.add(new float[fftComplexSize]);
-            segmentsIm.add(new float[fftComplexSize]);
+            segmentsRe.add(new double[fftComplexSize]);
+            segmentsIm.add(new double[fftComplexSize]);
         }
 
         for (int i = 0; i < segCount; i++) {
-            float[] re = new float[fftComplexSize];
-            float[] im = new float[fftComplexSize];
+            double[] re = new double[fftComplexSize];
+            double[] im = new double[fftComplexSize];
             int remaining = irLen - i * this.blockSize;
             int copySize = Math.max(0, Math.min(this.blockSize, remaining));
             Utilities.CopyAndPad(fftBuffer, ir, i * this.blockSize, copySize);
@@ -59,20 +59,20 @@ public class FFTConvolver {
             segmentsIRIm.add(im);
         }
 
-        preMultipliedRe = new float[fftComplexSize];
-        preMultipliedIm = new float[fftComplexSize];
-        convRe = new float[fftComplexSize];
-        convIm = new float[fftComplexSize];
-        overlap = new float[this.blockSize];
-        inputBuffer = new float[this.blockSize];
+        preMultipliedRe = new double[fftComplexSize];
+        preMultipliedIm = new double[fftComplexSize];
+        convRe = new double[fftComplexSize];
+        convIm = new double[fftComplexSize];
+        overlap = new double[this.blockSize];
+        inputBuffer = new double[this.blockSize];
         inputBufferFill = 0;
         current = 0;
         return true;
     }
 
-    public void process(float[] input, int inputOffset, float[] output, int outputOffset, int len) {
+    public void process(double[] input, int inputOffset, double[] output, int outputOffset, int len) {
         if (segCount == 0) {
-            Arrays.fill(output, outputOffset, outputOffset + len, 0f);
+            Arrays.fill(output, outputOffset, outputOffset + len, 0.0);
             return;
         }
         int processed = 0;
@@ -86,8 +86,8 @@ public class FFTConvolver {
             fft.fft(fftBuffer, segmentsRe.get(current), segmentsIm.get(current));
 
             if (inputBufferWasEmpty) {
-                Arrays.fill(preMultipliedRe, 0f);
-                Arrays.fill(preMultipliedIm, 0f);
+                Arrays.fill(preMultipliedRe, 0.0);
+                Arrays.fill(preMultipliedIm, 0.0);
                 for (int i = 1; i < segCount; i++) {
                     int indexIr = i;
                     int indexAudio = (current + i) % segCount;
@@ -114,7 +114,7 @@ public class FFTConvolver {
 
             inputBufferFill += processing;
             if (inputBufferFill == blockSize) {
-                Arrays.fill(inputBuffer, 0f);
+                Arrays.fill(inputBuffer, 0.0);
                 inputBufferFill = 0;
                 System.arraycopy(fftBuffer, blockSize, overlap, 0, blockSize);
                 current = current > 0 ? current - 1 : segCount - 1;
@@ -123,7 +123,7 @@ public class FFTConvolver {
         }
     }
 
-    public void process(float[] input, float[] output, int len) {
+    public void process(double[] input, double[] output, int len) {
         process(input, 0, output, 0, len);
     }
 
@@ -136,15 +136,15 @@ public class FFTConvolver {
         segmentsIm.clear();
         segmentsIRRe.clear();
         segmentsIRIm.clear();
-        fftBuffer = new float[0];
+        fftBuffer = new double[0];
         fft.init(0);
-        preMultipliedRe = new float[0];
-        preMultipliedIm = new float[0];
-        convRe = new float[0];
-        convIm = new float[0];
-        overlap = new float[0];
+        preMultipliedRe = new double[0];
+        preMultipliedIm = new double[0];
+        convRe = new double[0];
+        convIm = new double[0];
+        overlap = new double[0];
         current = 0;
-        inputBuffer = new float[0];
+        inputBuffer = new double[0];
         inputBufferFill = 0;
     }
 }
