+++ src/main/java/vvvfsimulator/audiofilter/TwoStageFFTConvolver.java
@@ -7,17 +7,17 @@ public class TwoStageFFTConvolver {
     private int tailBlockSize;
     private final FFTConvolver headConvolver = new FFTConvolver();
     private final FFTConvolver tailConvolver0 = new FFTConvolver();
-    private float[] tailOutput0 = new float[0];
-    private float[] tailPrecalculated0 = new float[0];
+    private double[] tailOutput0 = new double[0];
+    private double[] tailPrecalculated0 = new double[0];
     private final FFTConvolver tailConvolver = new FFTConvolver();
-    private float[] tailOutput = new float[0];
-    private float[] tailPrecalculated = new float[0];
-    private float[] tailInput = new float[0];
+    private double[] tailOutput = new double[0];
+    private double[] tailPrecalculated = new double[0];
+    private double[] tailInput = new double[0];
     private int tailInputFill;
     private int precalculatedPos;
-    private float[] backgroundProcessingInput = new float[0];
+    private double[] backgroundProcessingInput = new double[0];
 
-    public boolean init(int headBlockSize, int tailBlockSize, float[] ir, int irLen) {
+    public boolean init(int headBlockSize, int tailBlockSize, double[] ir, int irLen) {
         reset();
         if (headBlockSize == 0 || tailBlockSize == 0) {
             return false;
@@ -28,7 +28,7 @@ public class TwoStageFFTConvolver {
             headBlockSize = tailBlockSize;
             tailBlockSize = tmp;
         }
-        while (irLen > 0 && Math.abs(ir[irLen - 1]) < 0.000001f) {
+        while (irLen > 0 && Math.abs(ir[irLen - 1]) < 0.000001) {
             irLen--;
         }
         if (irLen == 0) {
@@ -42,28 +42,28 @@ public class TwoStageFFTConvolver {
 
         if (irLen > this.tailBlockSize) {
             int conv1IrLen = Math.min(irLen - this.tailBlockSize, this.tailBlockSize);
-            float[] tailIr0 = Arrays.copyOfRange(ir, this.tailBlockSize, this.tailBlockSize + conv1IrLen);
+            double[] tailIr0 = Arrays.copyOfRange(ir, this.tailBlockSize, this.tailBlockSize + conv1IrLen);
             tailConvolver0.init(this.headBlockSize, tailIr0, conv1IrLen);
-            tailOutput0 = new float[this.tailBlockSize];
-            tailPrecalculated0 = new float[this.tailBlockSize];
+            tailOutput0 = new double[this.tailBlockSize];
+            tailPrecalculated0 = new double[this.tailBlockSize];
         }
         if (irLen > 2 * this.tailBlockSize) {
             int tailIrLen = irLen - 2 * this.tailBlockSize;
-            float[] tailIr = Arrays.copyOfRange(ir, 2 * this.tailBlockSize, 2 * this.tailBlockSize + tailIrLen);
+            double[] tailIr = Arrays.copyOfRange(ir, 2 * this.tailBlockSize, 2 * this.tailBlockSize + tailIrLen);
             tailConvolver.init(this.tailBlockSize, tailIr, tailIrLen);
-            tailOutput = new float[this.tailBlockSize];
-            tailPrecalculated = new float[this.tailBlockSize];
-            backgroundProcessingInput = new float[this.tailBlockSize];
+            tailOutput = new double[this.tailBlockSize];
+            tailPrecalculated = new double[this.tailBlockSize];
+            backgroundProcessingInput = new double[this.tailBlockSize];
         }
         if (tailPrecalculated0.length > 0 || tailPrecalculated.length > 0) {
-            tailInput = new float[this.tailBlockSize];
+            tailInput = new double[this.tailBlockSize];
         }
         tailInputFill = 0;
         precalculatedPos = 0;
         return true;
     }
 
-    public void process(float[] input, int inputOffset, float[] output, int outputOffset, int len) {
+    public void process(double[] input, int inputOffset, double[] output, int outputOffset, int len) {
         headConvolver.process(input, inputOffset, output, outputOffset, len);
         if (tailInput.length == 0) {
             return;
@@ -94,7 +94,7 @@ public class TwoStageFFTConvolver {
                 int blockOffset = tailInputFill - headBlockSize;
                 tailConvolver0.process(tailInput, blockOffset, tailOutput0, blockOffset, headBlockSize);
                 if (tailInputFill == tailBlockSize) {
-                    float[] tmp = tailPrecalculated0;
+                    double[] tmp = tailPrecalculated0;
                     tailPrecalculated0 = tailOutput0;
                     tailOutput0 = tmp;
                 }
@@ -102,7 +102,7 @@ public class TwoStageFFTConvolver {
             if (tailPrecalculated.length > 0 && tailInputFill == tailBlockSize
                     && backgroundProcessingInput.length == tailBlockSize && tailOutput.length == tailBlockSize) {
                 waitForBackgroundProcessing();
-                float[] tmp = tailPrecalculated;
+                double[] tmp = tailPrecalculated;
                 tailPrecalculated = tailOutput;
                 tailOutput = tmp;
                 System.arraycopy(tailInput, 0, backgroundProcessingInput, 0, tailBlockSize);
@@ -118,7 +118,7 @@ public class TwoStageFFTConvolver {
         }
     }
 
-    public void process(float[] input, float[] output, int len) {
+    public void process(double[] input, double[] output, int len) {
         process(input, 0, output, 0, len);
     }
 
@@ -127,15 +127,15 @@ public class TwoStageFFTConvolver {
         tailBlockSize = 0;
         headConvolver.reset();
         tailConvolver0.reset();
-        tailOutput0 = new float[0];
-        tailPrecalculated0 = new float[0];
+        tailOutput0 = new double[0];
+        tailPrecalculated0 = new double[0];
         tailConvolver.reset();
-        tailOutput = new float[0];
-        tailPrecalculated = new float[0];
-        tailInput = new float[0];
+        tailOutput = new double[0];
+        tailPrecalculated = new double[0];
+        tailInput = new double[0];
         tailInputFill = 0;
         precalculatedPos = 0;
-        backgroundProcessingInput = new float[0];
+        backgroundProcessingInput = new double[0];
     }
 
     protected void startBackgroundProcessing() {
