+++ src/main/java/vvvfsimulator/generation/audio/trainsound/AudioFilter.java
@@ -1,32 +1,30 @@
 package vvvfsimulator.generation.audio.trainsound;
 
-import vvvfsimulator.audiofilter.AudioFFT;
 import vvvfsimulator.audiofilter.FFTConvolver;
-import vvvfsimulator.audiofilter.Utilities;
 
 public final class AudioFilter {
     private AudioFilter() {
     }
 
     public interface SampleFilter {
-        float apply(float input);
+        double apply(double input);
     }
 
     public static class IdentityFilter implements SampleFilter {
         @Override
-        public float apply(float input) {
+        public double apply(double input) {
             return input;
         }
     }
 
     public static final class CppConvolutionFilter {
         private final FFTConvolver convolver = new FFTConvolver();
-        private final float[] input;
-        private final float[] output;
+        private final double[] input;
+        private final double[] output;
 
-        public CppConvolutionFilter(int blockSize, float[] response) {
-            this.input = new float[blockSize];
-            this.output = new float[blockSize];
+        public CppConvolutionFilter(int blockSize, double[] response) {
+            this.input = new double[blockSize];
+            this.output = new double[blockSize];
             convolver.init(blockSize, response, response.length);
         }
 
@@ -34,11 +32,11 @@ public final class AudioFilter {
             convolver.reset();
         }
 
-        public void process(float[] in, int inOffset, float[] out, int outOffset, int len) {
+        public void process(double[] in, int inOffset, double[] out, int outOffset, int len) {
             convolver.process(in, inOffset, out, outOffset, len);
         }
 
-        public void process(float[] inOut, int len) {
+        public void process(double[] inOut, int len) {
             if (len > input.length) {
                 throw new IllegalArgumentException("len exceeds block size");
             }
@@ -47,7 +45,7 @@ public final class AudioFilter {
             System.arraycopy(output, 0, inOut, 0, len);
         }
 
-        public static void stereo2monaural(float[] input, int len, float[] outputL, float[] outputR) {
+        public static void stereo2monaural(double[] input, int len, double[] outputL, double[] outputR) {
             int frames = len / 2;
             for (int i = 0; i < frames; i++) {
                 outputL[i] = input[2 * i];
@@ -55,7 +53,7 @@ public final class AudioFilter {
             }
         }
 
-        public static void monaural2stereo(float[] inputL, float[] inputR, float[] output, int len) {
+        public static void monaural2stereo(double[] inputL, double[] inputR, double[] output, int len) {
             int frames = len / 2;
             for (int i = 0; i < frames; i++) {
                 output[2 * i] = inputL[i];
